+++ src/modules/test.js
@@ -1,0 +1,4 @@
+import { createMyServer, unknownObject } from "./cjsToEsm.mjs";
+
+console.log(unknownObject);
+console.log(createMyServer);
